Revert "File 294 integration tests"
